pyproject.toml
[project]
name = "classroom-repos"
version = "0.1.1"
description = "Manage shared files across local GitHub Classroom repositories."
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
dependencies = [
    "pyyaml>=6.0",
    "typer>=0.12",
]

[project.urls]
Homepage = "https://github.com/uwf-fang/gh-classroom-repos"
Repository = "https://github.com/uwf-fang/gh-classroom-repos"
Issues = "https://github.com/uwf-fang/gh-classroom-repos/issues"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.14",
]

[project.scripts]
classroom-repos = "classroom_repos.cli:app"

[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
